upscale to 16 instead of 32

+++ CUDA/GB_cuda_reduce_jitFactory.hpp
@@ -130,7 +130,7 @@ class reduceFactory
 
         // allocate and initialize zscalar
         size_t zsize = monoid->op->ztype->size ;
-        size_t zscalar_size = GB_IMAX (zsize, sizeof (int32_t)) ;
+        size_t zscalar_size = GB_IMAX (zsize, sizeof (uint16_t)) ;
         void *zscalar = rmm_wrap_malloc (zscalar_size) ;
         if (zscalar == NULL)
         {
